ScrapeEngine: Move ScrapeEngine to the Service namespace
