+++ Snowflake.Service/Snowflake.Service.csproj
@@ -91,6 +91,7 @@
     <Compile Include="Service\Manager\EmulatorAssembliesManager.cs" />
     <Compile Include="Service\Manager\PluginManager.cs" />
     <Compile Include="Service\Manager\ServerManager.cs" />
+    <Compile Include="Service\ScrapeEngine.cs" />
     <Compile Include="Service\Server\FileMediaStoreServer.cs" />
     <Compile Include="Service\Server\GameCacheServer.cs" />
     <Compile Include="Service\Server\JsonApiServer.cs" />
