+++ Snowflake/Snowflake.csproj
@@ -143,7 +143,6 @@
     <Compile Include="Platform\PlatformInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Scraper\ScrapableInfo.cs" />
-    <Compile Include="Scraper\ScrapeEngine.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
